requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.2
channels==2.4.0
channels-redis==3.1.0
constantly==15.1.0
cryptography==3.1
daphne==2.5.0
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.4.0
django-filter==2.3.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
iniconfig==1.0.1
install==1.3.1
msgpack==1.0.0
mysqlclient==2.0.1
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.1.0
pytest-asyncio==0.14.0
pytest-django==3.10.0
pytz==2020.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
whitenoise==5.2.0
zope.interface==5.1.0
